Add files via upload
This is a weather forecast model designed for three counties in Kenya: Nakuru, Narok, and Nyandarua. The purpose of this project is to develop a prototype model to assist small-scale potato farmers in these counties, where a significant portion of the population relies on potato farming as a primary livelihood. The model aims to provide valuable weather insights—such as precipitation and temperature forecasts—to support critical agricultural activities, including planting, irrigation, and harvesting decisions.